+++ devops/docker/run_docker.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+docker run --rm -i -t -p 7000:7000 nimwc:1.0    # Run docker.