+++ nimwcpkg/resources/administration/create_standarddata.nim
@@ -11,10 +11,7 @@ const
 
   head* = """
   <meta charset="utf-8">
-  <title>Nim Website Creator</title>
   <meta name="generator" content="Nim Website Creator">
-  <meta name="description" content="Nim Website Creator">
-  <meta name="keywords" content="Nim Website Creator">
   <meta name="viewport" content="width=device-width, initial-scale=1">
 
   <link rel="shortcut icon" href="/favicon.ico">
@@ -34,10 +31,7 @@ const
 
   headClean = """
   <meta charset="utf-8">
-  <title></title>
   <meta name="generator" content="">
-  <meta name="description" content="">
-  <meta name="keywords" content="">
   <meta name="viewport" content="width=device-width, initial-scale=1">
 
   <link rel="stylesheet" href="/css/style.css">
@@ -54,12 +48,12 @@ const
   <link rel="icon" type="image/png" href="/images/logo/favicon-32x32.png" sizes="32x32">
   <link rel="icon" type="image/png" href="/images/logo/favicon-192x192.png" sizes="192x192">
   <link rel="apple-touch-icon" sizes="180x180" href="/images/logo/favicon-180x180.png">
-  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha256-YLGeXaapI0/5IgZopewRJcFXomhRMlYYjugPLSyNjTY=" crossorigin="anonymous" />
   <link rel="stylesheet" href="/css/style.css">
   <link rel="stylesheet" href="/css/style_custom.css">
   <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha256-CjSoeELFOcH0/uxWu6mC/Vlrc1AARqbm/jiiImDGV3s=" crossorigin="anonymous" defer></script>
   <script src="/js/js.js" defer></script>
   <script src="/js/js_custom.js" defer></script>
   """
@@ -345,7 +339,7 @@ const
                 </h4>
                 <p style="margin-bottom: 0rem;">
                   <label style="width: 70px;">Clone:</label>
-                  <kbd>git clone https://git.io/f4AfL</kbd>
+                  <kbd>git clone https://github.com/ThomasTJdev/nim_websitecreator.git</kbd>
                 </p>
                 <p>
                   <label style="width: 70px;">Compile:</label>
@@ -385,7 +379,7 @@ const
                 <h4>Learn more
                 </h4>
                 <p style="margin-bottom: 0.3rem;">You can visit the
-                  <a href="https://git.io/f4AfL">Github page</a> to see the code
+                  <a href="https://github.com/ThomasTJdev/nim_websitecreator">Github page</a> to see the code
                 </p>
                 <p>Checkout the features and see the examples on how to use NimWC
                 </p>