+++ nimwcpkg/resources/files/files_efs.nim
@@ -3,36 +3,38 @@ import parsecfg, os, strutils, logging, ../utils/logging_nimwc
 
 setCurrentDir(getAppDir().replace("/nimwcpkg", ""))
 
-const
-  section = when defined(dev): "storagedev" else: "storage"
-  storageEFS* = "files/efs"
+const section = when defined(dev): "storagedev" else: "storage"
 
-let
-  appDir = replace(getAppDir(), "/nimwcpkg", "")
-  dict = loadConfig("config/config.cfg")
-  tempDir = dict.getSectionValue("Storage", section)
-  paths2create = [
-    "files", "files/efs", "fileslocal", tempDir & "/tmp",
-    tempDir & "/files", tempDir & "/files/private",
-    tempDir & "/files/public", tempDir & "/users",
+var storageEFS*: string
+
+
+proc setFolderPath(): string =
+  ## Sets the global folder path
+  let dict = loadConfig("config/config.cfg")
+  result = dict.getSectionValue("Storage", section)
+
+  if result == "fileslocal" or result == "":
+    result = replace(getAppDir(), "/nimwcpkg", "") / result
+
+  info("Storage path set to: " & result)
+  return result
+
+
+proc createFolders() =
+  ## Create folders
+  info("Checking that required folders exists.")
+  let paths2create = [
+    storageEFS,
+    storageEFS & "/tmp",
+    storageEFS & "/files",
+    storageEFS & "/files/private",
+    storageEFS & "/files/public",
+    storageEFS & "/users"
   ]
+  for folder in paths2create:
+    discard existsOrCreateDir(folder)
+    assert fpUserWrite in getFilePermissions(folder), "Wrong folder permissions:\n" & $getFilePermissions(folder) & folder
 
-# Create folders
-info("Checking that required 'files' folders exists.")
-for folder in paths2create:
-  discard existsOrCreateDir(folder)
-  assert fpUserWrite in getFilePermissions(folder), "Wrong folder permissions:\n" & $getFilePermissions(folder) & folder
 
-# Storage settings
-if tempDir == "fileslocal" or defined(dev):
-  info("Symlinking " & tempDir & " to files/efs")
-  try:
-    createSymlink(src= appDir & "/" & tempDir & "/*",
-                  dest=appDir & "/files/efs/")
-  except: discard
-else:
-  info("Symlinking " & tempDir & " to files/efs")
-  try:
-    createSymlink(src= tempDir & "/*",
-                  dest=appDir & "/files/efs/")
-  except: discard
+storageEFS = setFolderPath()
+createFolders()