+++ nimwcpkg/resources/web/routes.nim
@@ -544,7 +544,7 @@ routes:
     if @"access" notin ["private", "public", "publicimage"]:
       resp("Error: Missing access right")
 
-    const
+    let
       efspublic = storageEFS & "/files/public/"
       efsprivate = storageEFS & "/files/private/"
     var
@@ -558,8 +558,8 @@ routes:
     if not usesWebp and @"checksum" == "true":
       filename = getMD5(filedata) & fileexts
 
-    if @"normalize" == "true":
-      filename = filename.normalize
+    if @"lowercase" == "true":
+      filename = filename.toLowerAscii()
 
     if @"access" == "publicimage":
       path = "public/images/" & filename
@@ -789,7 +789,7 @@ routes:
     ## Get a file
     createTFD()
     let filename = decodeUrl(@"filename")
-    var filepath = storageEFS & "/users/" & filename
+    let filepath = storageEFS & "/users/" & filename
     if not fileExists(filepath): resp("")
     sendFile(filepath)
 
@@ -838,7 +838,7 @@ routes:
 
     let blogID = insertID(db, sql"INSERT INTO blog (author_id, status, url, name, description, standardhead, standardnavbar, standardfooter, title, metadescription, metakeywords, category, tags, pubDate, viewCount) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)", c.userid, @"status", url, @"name", @"editordata", checkboxToInt(@"standardhead"), checkboxToInt(@"standardnavbar"), checkboxToInt(@"standardfooter"), @"title", @"metadescription", @"metakeywords", @"category", @"tags", @"pubdate", @"viewcount")
 
-    resp genMainAdmin(c, genEditBlog(c, $blogID, true), "edit")
+    redirect("/editpage/blog/" & $blogID & "?newpage=true")
 
 
   post "/blogpage/update":
@@ -877,7 +877,7 @@ routes:
     createTFD()
     restrictAccessTo(c, [Admin, Moderator])
 
-    resp genMainAdmin(c, genEditBlog(c, @"blogid"), "edit")
+    resp genMainAdmin(c, genEditBlog(c, @"blogid", @"newpage"), "edit")
 
 
   get "/blog":
@@ -919,7 +919,7 @@ routes:
 
     let pageID = insertID(db, sql"INSERT INTO pages (author_id, status, url, name, description, standardhead, standardnavbar, standardfooter, title, metadescription, metakeywords, category, tags) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)", c.userid, @"status", url, @"name", @"editordata", checkboxToInt(@"standardhead"), checkboxToInt(@"standardnavbar"), checkboxToInt(@"standardfooter"), @"title", @"metadescription", @"metakeywords", @"category", @"tags")
 
-    resp genMainAdmin(c, genEditPage(c, $pageID, true), "edit")
+    redirect("/editpage/page/" & $pageID & "?newpage=true")
 
 
   post "/page/update":
@@ -960,7 +960,7 @@ routes:
     createTFD()
     restrictAccessTo(c, [Admin, Moderator])
 
-    resp genMainAdmin(c, genEditPage(c, @"pageid"), "edit")
+    resp genMainAdmin(c, genEditPage(c, @"pageid", @"newpage"), "edit")
 
 
   get re"/p//*.":