+++ public/js/js_admin.js
@@ -68,14 +68,14 @@ if ($btnFileDelete.length) {
 function uploadFile(projectID) {
   var access    = $("input[name=fileRadio]:checked").attr("data-value");
   var webp      = $("#webpstatus").is(":checked");
-  var norm      = $("#normalize").is(":checked");
+  var norm      = $("#lowercase").is(":checked");
   var chks      = $("#checksum").is(":checked");
   var file      = $("#file").get(0).files[0];
   var filename  = file.name;
   var formData  = new FormData();
   formData.append('file', file);
   $.ajax({
-    url: "/files/upload/" + access + "?webpstatus=" + webp + "&normalize=" + norm + "&checksum=" + chks,
+    url: "/files/upload/" + access + "?webpstatus=" + webp + "&lowercase=" + norm + "&checksum=" + chks,
     data: formData,
     type: 'POST',
     cache: false,